requirements.txt
flask>=3.1.2
flask-cors>=4.0.0
transformers>=4.57.1
torch>=2.9.0
nltk>=3.9.2
pandas>=2.3.3
numpy>=2.3.4
scikit-learn>=1.7.2
python-dotenv>=1.0.0
aiohttp>=3.9.0
sentencepiece  # Required for M2M100 model
textblob>=0.18.0
google-generativeai>=0.4.0